use flaot

+++ src/main/java/team/cqr/cqrepoured/entity/profile/variant/DamageEntry.java
@@ -11,14 +11,14 @@ import net.minecraft.resources.ResourceLocation;
 
 public record DamageEntry(
 		Optional<DamageCap> damageCap,
-		double minDamage,
+		float minDamage,
 		boolean fireImmune,
 		Map<ResourceLocation, Float> damageTypeMultipliers
 ) {
 	public static final Codec<DamageEntry> CODEC = RecordCodecBuilder.create(instance -> {
 		return instance.group(
 				DamageCap.CODEC.optionalFieldOf("damage-cap").forGetter(DamageEntry::damageCap),
-				Codec.DOUBLE.fieldOf("min-damage").forGetter(DamageEntry::minDamage),
+				Codec.FLOAT.fieldOf("min-damage").forGetter(DamageEntry::minDamage),
 				Codec.BOOL.fieldOf("fire-immune").forGetter(DamageEntry::fireImmune),
 				Codec.unboundedMap(ResourceLocation.CODEC, Codec.FLOAT).optionalFieldOf("multipliers", Collections.emptyMap()).forGetter(DamageEntry::damageTypeMultipliers)
 			).apply(instance, DamageEntry::new);
